using azure for transcription
