+++ package.json
@@ -12,7 +12,10 @@
   },
   "dependencies": {
     "@logtail/node": "^0.5.2",
+    "@radix-ui/react-scroll-area": "^1.2.3",
     "@sentry/nextjs": "^9.3.0",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
     "dotenv": "^16.4.7",
     "lucide-react": "^0.484.0",
     "microsoft-cognitiveservices-speech-sdk": "^1.43.0",
@@ -20,6 +23,7 @@
     "next": "15.1.6",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "tailwind-merge": "^3.0.2",
     "winston": "^3.17.0"
   },
   "devDependencies": {
