+++ src/app/transcription/page.tsx
@@ -1,0 +1,7 @@
+import { TranscriptionUI } from "@/components/ui/transcription";
+
+const TranscriptionPage = () => {
+  return <TranscriptionUI />;
+};
+
+export default TranscriptionPage;
